requirements.txt
streamlit==1.40.2
pandas==2.2.3
pillow==11.0.0
python-dotenv==1.0.1
langchain==0.3.9
langchain-openai==0.2.10
openai==1.55.3